+++ app/connectors_service/connectors/sources/postgresql/__init__.py
@@ -1,0 +1,17 @@
+#
+# Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+# or more contributor license agreements. Licensed under the Elastic License 2.0;
+# you may not use this file except in compliance with the Elastic License 2.0.
+#
+
+from .client import PostgreSQLClient
+from .datasource import PostgreSQLDataSource
+from .queries import PostgreSQLQueries
+from .validator import PostgreSQLAdvancedRulesValidator
+
+__all__ = [
+    "PostgreSQLDataSource",
+    "PostgreSQLClient",
+    "PostgreSQLAdvancedRulesValidator",
+    "PostgreSQLQueries",
+]
